+++ src/main/java/com/brand/blockus/content/types/WoodTypes.java
@@ -46,11 +46,13 @@ public class WoodTypes {
         FabricBlockSettings blockSettings = FabricBlockSettings.create().mapColor(color).instrument(Instrument.BASS).strength(2.0F, 3.0F).sounds(sound).burnable();
         FabricBlockSettings blockSettings2 = BlocksRegistration.createDoorTrapdoorBlockSettings(0.1f, 0.8f, sound, color, Instrument.BASS).burnable();
         FabricBlockSettings signSettings = FabricBlockSettings.create().mapColor(color).noCollision().strength(1.0F).sounds(sound);
+        FabricBlockSettings hangingSignSettings = FabricBlockSettings.create().mapColor(color).noCollision().strength(1.0F).sounds(sound);
 
         if (burnable) {
             blockSettings = blockSettings.burnable();
             blockSettings2 = blockSettings2.burnable();
             signSettings = signSettings.burnable();
+            hangingSignSettings = hangingSignSettings.burnable();
         }
 
         this.planks = BlocksRegistration.register(type + "_planks", new Block(blockSettings));
@@ -71,8 +73,8 @@ public class WoodTypes {
 
         Identifier hangingSignPath = new Identifier(Blockus.MOD_ID, "entity/signs/hanging/" + type);
         Identifier hangingSignGuiPath = new Identifier(Blockus.MOD_ID, "textures/gui/hanging_signs/" + type);
-        this.ceiling_hanging_sign = BlocksRegistration.registerNoItem(type + "_hanging_sign", new TerraformHangingSignBlock(hangingSignPath, hangingSignGuiPath, signSettings));
-        this.wall_hanging_sign = BlocksRegistration.registerNoItem(type + "_wall_hanging_sign", new TerraformWallHangingSignBlock(hangingSignPath, hangingSignGuiPath, signSettings.dropsLike(ceiling_hanging_sign)));
+        this.ceiling_hanging_sign = BlocksRegistration.registerNoItem(type + "_hanging_sign", new TerraformHangingSignBlock(hangingSignPath, hangingSignGuiPath, hangingSignSettings));
+        this.wall_hanging_sign = BlocksRegistration.registerNoItem(type + "_wall_hanging_sign", new TerraformWallHangingSignBlock(hangingSignPath, hangingSignGuiPath, hangingSignSettings.dropsLike(ceiling_hanging_sign)));
         this.hanging_sign = BlockusItems.registerHangingSign(ceiling_hanging_sign, wall_hanging_sign);
 
         LIST.add(this);