+++ changelog.md
@@ -1,2 +1,2 @@
-* Fixed wrong Soul Sandstone, Cut Soul Sandstone & Chiseled Soul Sandstone top & bottom textures
+* Fixed signs & hanging signs drop